Reorganize packet package structure
