+++ src/main/java/kernitus/plugin/OldCombatMechanics/module/ModuleAttackSounds.java
@@ -3,9 +3,9 @@ package kernitus.plugin.OldCombatMechanics.module;
 import kernitus.plugin.OldCombatMechanics.OCMMain;
 import kernitus.plugin.OldCombatMechanics.utilities.Config;
 import kernitus.plugin.OldCombatMechanics.utilities.Messenger;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketAdapter;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketEvent;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketManager;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketAdapter;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketEvent;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketManager;
 import kernitus.plugin.OldCombatMechanics.utilities.packet.sound.SoundPacket;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
