+++ src/main/java/kernitus/plugin/OldCombatMechanics/module/ModuleSwordSweep.java
@@ -3,9 +3,9 @@ package kernitus.plugin.OldCombatMechanics.module;
 import kernitus.plugin.OldCombatMechanics.OCMMain;
 import kernitus.plugin.OldCombatMechanics.utilities.Messenger;
 import kernitus.plugin.OldCombatMechanics.utilities.damage.ToolDamage;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketAdapter;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketEvent;
-import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketManager;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketAdapter;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketEvent;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.mitm.PacketManager;
 import kernitus.plugin.OldCombatMechanics.utilities.packet.particle.ParticlePacket;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
