+++ src/main/java/kernitus/plugin/OldCombatMechanics/utilities/packet/mitm/PacketManager.java
@@ -1,4 +1,4 @@
-package kernitus.plugin.OldCombatMechanics.utilities.packet;
+package kernitus.plugin.OldCombatMechanics.utilities.packet.mitm;
 
 import kernitus.plugin.OldCombatMechanics.OCMMain;
 import kernitus.plugin.OldCombatMechanics.utilities.Messenger;
