+++ src/main/java/kernitus/plugin/OldCombatMechanics/utilities/packet/particle/ParticlePacket.java
@@ -2,8 +2,8 @@ package kernitus.plugin.OldCombatMechanics.utilities.packet.particle;
 
 import kernitus.plugin.OldCombatMechanics.utilities.packet.ImmutablePacket;
 import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketHelper;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketHelper.PacketType;
 import kernitus.plugin.OldCombatMechanics.utilities.reflection.Reflector;
-import kernitus.plugin.OldCombatMechanics.utilities.reflection.type.PacketType;
 
 import java.util.Optional;
