+++ src/main/java/kernitus/plugin/OldCombatMechanics/utilities/packet/team/PreV17TeamPacket.java
@@ -1,8 +1,8 @@
 package kernitus.plugin.OldCombatMechanics.utilities.packet.team;
 
 import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketHelper;
+import kernitus.plugin.OldCombatMechanics.utilities.packet.PacketHelper.PacketType;
 import kernitus.plugin.OldCombatMechanics.utilities.reflection.Reflector;
-import kernitus.plugin.OldCombatMechanics.utilities.reflection.type.PacketType;
 import kernitus.plugin.OldCombatMechanics.utilities.teams.CollisionRule;
 import kernitus.plugin.OldCombatMechanics.utilities.teams.TeamAction;
 import org.bukkit.entity.Player;
