--- src/main/java/kernitus/plugin/OldCombatMechanics/utilities/reflection/type/PacketType.java
@@ -1,13 +1,0 @@
-package kernitus.plugin.OldCombatMechanics.utilities.reflection.type;
-
-public enum PacketType {
-
-    PlayOut("PlayOut"), PlayIn("PlayIn");
-
-    public String prefix;
-
-    PacketType(String prefix){
-        this.prefix = prefix;
-    }
-
-}
